added chatgpt, cmd line arg parsing and refactored map rendering methods
