--- basic_pipeline/world_gen/.env
@@ -1,1 +1,0 @@
-OPENAI_API_KEY=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
