--- basic_pipeline/world_gen/config.json
@@ -1,11 +1,0 @@
-{
-    "size": "XL",
-    "seed": 42,
-    "biomes": {"north": "plains",
-                    "northeast": "mountains",
-                    "east": "plains",
-                    "west": "desert",
-                    "northwest": "plains",
-                    "southwest": "water",
-                    "center": "plains"}
-}
