+++ basic_pipeline/world_gen/gpt_api.py
@@ -5,7 +5,7 @@ from dotenv import load_dotenv
 import os
 
 # Load API key from environment variables
-load_dotenv()
+load_dotenv(dotenv_path="../api/.env")
 api_key = os.getenv("OPENAI_API_KEY")
 
 def extract_world_data(prompt):
@@ -13,7 +13,7 @@ def extract_world_data(prompt):
     system_message = (
         "You are an AI assistant that extracts world generation parameters from natural language prompts and returns them as a JSON object.\n"
         "Ensure the response is a valid JSON object with the following fields:\n"
-        "- 'biomes': a dictionary mapping 'north', 'south', 'east', 'west', 'northeast', 'southeast', 'northwest', 'southwest' to biomes ('water', 'desert', 'plains', 'forest', 'mountains').\n"
+        "- 'biomes': a dictionary mapping 'north', 'south', 'east', 'west', 'northeast', 'southeast', 'northwest', 'southwest' and 'center' to biomes ('water', 'desert', 'plains', 'forest', 'mountains').\n"
         "- 'temperature': a dictionary mapping regions to temperature descriptions.\n"
         "- 'precipitation': a dictionary mapping regions to precipitation descriptions.\n"
         "- 'seed': an optional alphanumeric string if the user specifies one.\n"
@@ -34,7 +34,6 @@ def extract_world_data(prompt):
         )
 
         extracted_data = response.choices[0].message.content.strip()  # Get response text
-        print(extracted_data)  # Print raw output for debugging
 
         return json.loads(extracted_data)  # Convert to JSON
     except Exception as e:
@@ -44,6 +43,7 @@ def main():
     parser = argparse.ArgumentParser(description="Extract world parameters from a prompt.")
     parser.add_argument("input_text", type=str, help="User description of the world")
     args = parser.parse_args()
+    print(args)
 
     world_data = extract_world_data(args.input_text)
     print("\nExtracted JSON:\n", json.dumps(world_data, indent=4))
