--- basic_pipeline/world_gen/one-off-files/b_mask.txt
@@ -1,36 +1,0 @@
-Creating a biome mask
-Displaying biome mask
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-000000000000000000000000000000000
-None
