--- basic_pipeline/world_gen/one-off-files/rainbow_text.py
@@ -1,13 +1,0 @@
-# Rainbow Text
-# "\033[Mode, Color Format (2 = RGB), Red, Green, Blue]mText To Print\033[0m"]
-print("\033[38;2;255;0;255mPink text\033[0m") 
-print("\033[38;5;196mBright red text\033[0m") 
-print("\033[38;2;255;100;0mOrange text\033[0m") 
-print("\033[38;2;255;210;0mYellow text\033[0m") 
-print("\033[38;2;150;230;0mLime green text\033[0m") 
-print("\033[38;2;0;255;0mGreen text\033[0m") 
-print("\033[38;2;64;244;208mTurquoise text\033[0m") 
-print("\033[38;2;20;110;100mTeal text\033[0m") 
-print("\033[38;2;0;100;255mBlue text\033[0m") 
-print("\033[38;2;0;0;255mIndigo text\033[0m")
-print("\033[38;2;100;50;255mViolet text\033[0m")
