--- basic_pipeline/world_gen/one-off-files/temp_map.txt
@@ -1,35 +1,0 @@
-Intel MKL WARNING: Support of Intel(R) Streaming SIMD Extensions 4.2 (Intel(R) SSE4.2) enabled only processors has been deprecated. Intel oneAPI Math Kernel Library 2025.0 will require Intel(R) Advanced Vector Extensions (Intel(R) AVX) instructions.
-Intel MKL WARNING: Support of Intel(R) Streaming SIMD Extensions 4.2 (Intel(R) SSE4.2) enabled only processors has been deprecated. Intel oneAPI Math Kernel Library 2025.0 will require Intel(R) Advanced Vector Extensions (Intel(R) AVX) instructions.
-n..~....n..~.~~~~~~~~~~~~~~~~~~~n
-.n......nnn....~~~~~~~~~~.~~~~~~.
-..nnn.nnnnnnn...~~~~~~~~~~.~~~~~~
-~..nnnnnnnnnnn...~~~~~~~~~~~~~~~~
-~...nnnnnnnnnnn...~~~~~~~~~~~~~~~
-~....nnnnnnnnnn...~~~~~~~~~~~~~~~
-~~...nnnnnnnnnnn...~~~~~~~~~~~~~~
-~~~..nnnnnnnnnnnn...~~~~~~~~~~~~~
-~~...nnnnnnnnnnnn...~~~~~~~~~~~~~
-~~....nnnnnnnnnnnn...~~~~~~~~~~~~
-~~~....nnnnnnnnnnnn...~~~~~~~~~~~
-~~~~....nnnnnnnnnnnn...~~~~~~~~~~
-~~~~.......nnnnnnnn.n...~~~~~~~~~
-~~~~~~~.....nnnnnn......~~~~~~~~~
-~~~~~~~~....nnnnnnn.n...~~~~~~~~~
-~~~~~~~~~.....nnnnnnn...~~~~~~~~~
-~~~~~~~~~~~...nnnnnnnn..~~~~~~~~~
-~~~~~~~~~~~~....nnnnnnn..~~~~~~~~
-~~~~~~~~~~~~.....nnnnnn...~~~~~~~
-~~~~~~~~~~~~~.....nnnnn.....~~~~~
-~~~~~~~~~~~~~~~....nnnn......~~~~
-~~~~~~~~~~~~~~~~~...nnnn.....~~~~
-~~~~~~~~~~~~~~~~~~...nnn......~~~
-~~~~~~~~~~~~~~~~~~~....n........~
-~~~~~~~~~~~~~~~~~~~~.............
-~~~~~~~~~~~~~~~~~~~~~~...n......~
-~~~~~~~~~~~~~~~~~~~~~~~...n.....~
-~~~~~~~~~~~~~~~~~~~~~~~~........~
-~~~~~~~~~~~~~~~~~~~~~~~~~........
-~~~~~~~~~~~~~~~~~~~~~~~~~~.......
-~~~~~~~~~~~~~~~~~~~~~~~~~~~......
-~~~~~~~.~~~~~~~~~~~~~~~~~~~~.....
-n.~~~~~..~~~~~~~~~~~~~~~~~~~~~~.n
