--- basic_pipeline/world_gen/sample_inputs.txt
@@ -1,21 +1,0 @@
-user_params = {'north': 'desert',
-                'south': 'mountains',
-                'center': 'forest'}
-                
-user_params = {'north': 'water',
-                'south': 'mountains',
-                'center': 'water'}
-
-user_params = {'north': 'plains',
-                'northeast': 'mountains',
-                'east': 'plains',
-                'west': 'desert',
-                "northwest": 'plains',
-                "southwest": 'water',
-                'center': 'plains'}
-
-user_params = {'north': 'water',
-                'south': 'water',
-                'east': 'water',
-                'west': 'water',
-                'center': 'mountains'}
