+++ basic_pipeline/world_gen/test_methods.py
@@ -1,0 +1,17 @@
+def generate_many_worlds(num_worlds, map_size, display_mode, generate_image=True):
+    '''
+    Use this to generate about n = num_worlds silently without the world 
+    generator displaying anything in Python.
+    
+    Desired I/O
+    Input: - This function will ouput the number of maps provided with 
+            the size and format specified.
+    Output: A directory will be made (if one is not specifed) that 
+            contains images of all of the worlds created and a text file that 
+            labels the JSON input used for that world. 
+    
+    Note: Images and JSON entries will be labeled numerically.
+    
+    ! I might move this to an evaluation method file...
+    '''
+    pass
